add userlogin resource and jwt_required to delete() of items and get claims

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
 
 
@@ -16,7 +16,6 @@ class Item(Resource):
                         help="Every item needs a store_id."
                         )
 
-    @jwt_required()
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -38,7 +37,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'you need admin priviledges'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -61,5 +65,6 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required
     def get(self):
-        return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
+        return {'items': [item.json() for item in ItemModel.find_all()]}
